add ally opponent nearby calculation

+++ src/MinionStrategy/RightAssociativeBinaryExpression.java
@@ -12,7 +12,7 @@ public class RightAssociativeBinaryExpression implements Expression {
     }
 
     @Override
-    public int evaluate(Minion minion) throws Exception {
+    public long evaluate(Minion minion) throws Exception {
         return  (int) Math.pow(left.evaluate(minion), right.evaluate(minion));
     }
